requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.6
chardet==3.0.4
confusable-homoglyphs==3.2.0
crispy-bootstrap5==0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.1
django-allauth==0.45.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-heroku==0.3.1
django-quill-editor==0.1.22
django-registration==3.2
django-rest-auth==0.9.5
django-suit==0.2.28
django-webpack-loader==1.1.0
django-wordpress-api==0.2.0
djangorestframework==3.12.2
gunicorn==20.1.0
idna==2.8
iso8601==0.1.12
Markdown==3.3.4
oauthlib==3.1.1
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
python3-openid==3.2.0
pytz==2020.1
requests==2.21.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.12.0
sqlparse==0.3.1
urllib3==1.24.3
waitress==2.0.0
whitenoise==5.2.0
